requirements.txt
escapejson
httpdis>=0.6.1
Mako
PyYAML>=3.10
pyinotify
redis>=2.4.0
requests>=2.0
six>=1.4.0
sonicprobe>=0.3.2
